Рефакторинг кода для взаимодействия с другими частями проекта.
