--- Located.cpp
@@ -1,15 +1,0 @@
-#include "Located.h"
-
-coord::coord(int x, int y) : x(x), y(y) {}
-
-coord::coord() : x(0), y(0) {}
-
-Located::Located(coord coordinates) : coordinates(coordinates) {}
-
-void Located::SetCoord(coord coordinates) {
-    this->coordinates = coordinates;
-}
-
-coord Located::GetCoord() const {
-    return coordinates;
-}
