+++ libAdapter.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <thread>
 
-#include "Movable.h"
+#include "movable.h"
 
 
 int Game();
